Fix the badly merged moved files
